Improve the Shared\Notifications\Messages
